cleaned up IS_JEST conditional
